requirements.txt
requests
sqlalchemy
python-telegram-bot
psycopg2
yandex-checkout